+++ src/vector.c
@@ -1,4 +1,4 @@
-#include "vector.h"
+#include "data_structures/vector.h"
 
 
 //===========================================================================//
